+++ CHANGELOG.rst
@@ -27,6 +27,8 @@ for more detailed information about upgrading to this release.
 * Update the ``chalice package`` command to support
   pure lambda functions and scheduled events.
   (`#772 <https://github.com/aws/chalice/issues/772>`__)
+* Fix packager edge case normalizing sdist names
+  (`#778 <https://github.com/aws/chalice/issues/778>`__)
 
 
 1.1.1